4.21*: bloglist expansion, step9

+++ part4/bloglist/controllers/blogs.js
@@ -66,8 +66,24 @@ blogsRouter.put('/:id', async (req, res) => {
 
 // DELETE BLOG
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndRemove(req.params.id)
-  res.status(204).end()
+  const body = req.body
+  
+  // check validity and decode token:
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid '} )
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(req.params.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(req.params.id)
+    res.status(204).end()
+  } else {
+    return res.status(401).json({ error: `cannot delete another user's blog` })
+  }
+  
 })
 
 module.exports = blogsRouter
